Move Vec4 to its own file, remove dependency on GLM
